mergning to main

+++ L01/ft_printf/libft/ft_putnbr_base.c
@@ -6,13 +6,13 @@
 /*   By: mkakizak <mkakizak@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/09 13:23:49 by mkakizak          #+#    #+#             */
-/*   Updated: 2024/05/11 17:45:51 by mkakizak         ###   ########.fr       */
+/*   Updated: 2024/05/11 18:16:37 by mkakizak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static void	recursive_print(long long num, int *res, char *base, unsigned long base_l)
+static void	recursive_print(unsigned long num, unsigned int *res, char *base, unsigned long base_l)
 {
 	char	print;
 
@@ -20,7 +20,6 @@ static void	recursive_print(long long num, int *res, char *base, unsigned long b
 	{
 		recursive_print (num / base_l, res, base , base_l);
 		print = base[num % base_l];
-		// printf("%c", print)
 ;		write(1, &print, 1);
 		*res += 1;
 	}
@@ -35,7 +34,6 @@ static void	recursive_print(long long num, int *res, char *base, unsigned long b
 		else
 		{
 			print = base[num % base_l];
-			// printf("printf is %c", print)
 			write(1, &print, 1);
 			*res += 1;
 		}
@@ -44,12 +42,19 @@ static void	recursive_print(long long num, int *res, char *base, unsigned long b
 
 int		ft_putnbr_base(unsigned long num, char *base)
 {
-	long  base_l;
+	unsigned long  base_l;
 	base_l = (unsigned long)ft_strlen(base);
-	int res;
+	unsigned int res;
 
 	res = 0;
 		recursive_print(num, &res, base, base_l);
 	return (res);
 	
+	
 }
+
+// int main(void)
+// {
+// 	printf("\nreturn value is %d\n", ft_putnbr_base(12345678, "0123456789abcdef"));
+// 	return(0);
+// }
